[enhancement](merge-on-write) add skip_delete_bitmap session variable for debug purpose

+++ be/src/vec/exec/scan/new_olap_scanner.cpp
@@ -260,7 +260,7 @@ Status NewOlapScanner::_init_tablet_reader_params(
         _tablet_reader_params.use_page_cache = true;
     }
 
-    if (_tablet->enable_unique_key_merge_on_write()) {
+    if (_tablet->enable_unique_key_merge_on_write() && !_state->skip_delete_bitmap()) {
         _tablet_reader_params.delete_bitmap = &_tablet->tablet_meta()->delete_bitmap();
     }
